New Tool - Logging
Logging tool added to Initializer Class. Can be started as Basic Initializing. When no operation is done in Basic for 10 seconds, the passivity log is recorded. When the motion is detected, the activity is recorded this time. These records are transferred into "Your Project Directory / Debug / Logs / Log.json" in Basic.
The GetLogs command has been added to the system to retrieve logs. Acceptable logs will be retrieved and examined by the Clients.